Migrate the inspector fake to null safety
